+++ packages/react-widgets/src/widgets/TableWidget.js
@@ -27,6 +27,9 @@ import { _FeatureFlags, _hasFeatureFlag } from '@carto/react-core';
  * @param  {boolean=} [props.stableHeight] -  If specified, error and no-data state will maintain same height as normal widget in loading/loaded state.
  * @param  {boolean} [props.dense] - Whether the table should use a compact layout with smaller cell paddings.
  * @param  {number} [props.pageSize] - Number of rows per page. This is used to manage internal state externally.
+ * @param  {string} [props.searchText] - Text to search in the table
+ * @param  {string} [props.searchColumn] - Column used to perform the search, using the searchText property
+ * @param  {Function} [props.search] - React component to handle search functionality
 
  */
 function TableWidget({
@@ -46,6 +49,9 @@ function TableWidget({
   height,
   stableHeight,
   dense,
+  searchText,
+  searchColumn,
+  search,
   // Internal state
   pageSize
 }) {
@@ -55,6 +61,7 @@ function TableWidget({
   const [sortBy, setSortBy] = useState(undefined);
   const [sortByColumnType, setSortByColumnType] = useState(undefined);
   const [sortDirection, setSortDirection] = useState('asc');
+  const containsStringSearchFilter = searchColumn && searchText;
 
   const {
     data = { rows: [], totalCount: 0, hasData: false },
@@ -65,6 +72,9 @@ function TableWidget({
     dataSource,
     params: {
       columns: [...columns.map((c) => c.field), ...hiddenColumnFields],
+      searchFilter: containsStringSearchFilter && {
+        [searchColumn]: { stringSearch: { values: [searchText] } }
+      },
       sortBy,
       sortDirection,
       sortByColumnType,
@@ -153,6 +163,9 @@ TableWidget.propTypes = {
   height: PropTypes.string,
   stableHeight: PropTypes.bool,
   dense: PropTypes.bool,
+  search: PropTypes.func,
+  searchText: PropTypes.string,
+  searchColumn: PropTypes.string,
   // Internal state
   pageSize: PropTypes.number
 };